% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{annualAgg}
\alias{annualAgg}
\alias{annualAgg.flowLoad}
\alias{annualAgg.rainLoad}
\alias{annualAgg.stageLoad}
\alias{annualAgg.rainAll}
\title{annualAgg}
\usage{
annualAgg(x, method = "mean", ...)

\method{annualAgg}{flowLoad}(x, method = "mean", ...)

\method{annualAgg}{rainLoad}(x, method = "mean", ...)

\method{annualAgg}{stageLoad}(x, method = "mean", ...)

\method{annualAgg}{rainAll}(x, method = "mean", ...)
}
\arguments{
\item{x}{Data generated in the HydroEnR package}

\item{method}{'mean', 'median', 'max', 'min', or 'sum'}

\item{...}{Other variables as required}
}
\value{

}
\description{
Aggregates sub annual time series into an annual resolution,
aggregations are carried out to the calendar day not a rolling time period.
}
\examples{
annualAgg(Buildwas)
}
