% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CD-class.R
\docType{class}
\name{CD-class}
\alias{CD-class}
\title{S4 class, \code{"CD"},  for catchment descriptor files}
\value{

}
\description{
S4 class, \code{"CD"},  for catchment descriptor files
}
\section{Slots}{

\describe{
\item{\code{ID}}{numeric.}

\item{\code{NAME}}{character.}

\item{\code{EASTING}}{numeric.}

\item{\code{NORTHING}}{numeric.}

\item{\code{FEHPOOL}}{character.}

\item{\code{FEHQMED}}{character.}

\item{\code{ALTBAR}}{numeric.}

\item{\code{ASPBAR}}{numeric.}

\item{\code{ASPVAR}}{numeric.}

\item{\code{AREA}}{numeric.}

\item{\code{BFIHOST}}{numeric.}

\item{\code{BFIHOST19}}{numeric.}

\item{\code{CENTROID}}{numeric.}

\item{\code{DDF}}{numeric.}

\item{\code{DPLBAR}}{numeric.}

\item{\code{DPSBAR}}{numeric.}

\item{\code{FARL}}{numeric.}

\item{\code{FPDBAR}}{numeric.}

\item{\code{FPEXT}}{numeric.}

\item{\code{FPLOC}}{numeric.}

\item{\code{LDP}}{numeric.}

\item{\code{PROPWET}}{numeric.}

\item{\code{RMED1H}}{numeric.}

\item{\code{RMED1D}}{numeric.}

\item{\code{RMED2D}}{numeric.}

\item{\code{SAAR}}{numeric.}

\item{\code{SAAR4170}}{numeric.}

\item{\code{SPRHOST}}{numeric.}

\item{\code{URBEXT1990}}{numeric.}

\item{\code{URBEXT2000}}{numeric.}

\item{\code{URBCON1990}}{numeric.}

\item{\code{URBCON2000}}{numeric.}

\item{\code{URBLOC1990}}{numeric.}

\item{\code{URBLOC2000}}{numeric.}
}}

