% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvFix.R
\name{csvFix}
\alias{csvFix}
\title{csvFix}
\usage{
csvFix(
  file,
  new.name = TRUE,
  sep = "|",
  comment.char = "",
  rowskip = 0,
  expected_col = 5,
  edit = FALSE
)
}
\arguments{
\item{file}{File link}

\item{new.name}{Set as TRUE, you will require to give fixed file a new name}

\item{sep}{Delimiter used in file}

\item{comment.char}{Include if comments are differentiated}

\item{rowskip}{Set as 0, but can offset where the function starts}

\item{expected_col}{Constrains the csv to a certain number of columns}

\item{edit}{If you wish to manually change file contents set as TRUE}
}
\value{

}
\description{
Data preprocessing for WISKI files, Constrains an irregular csv
columns to those of a user specified value. This negates the issue where
commented data have commas included.

Adapted from https://gist.github.com/noamross/4114258
}
\examples{
#csvFix("C:/Users/jpayne05/Downloads/Wellingborough.RE[RainfallEvent].15min.csv.all")
}
