% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumSkipNA.R
\name{cumsumNA}
\alias{cumsumNA}
\alias{cumsumNA.numeric}
\alias{cumsumNA.rainAll}
\alias{cumsumNA.rainAllDaily}
\alias{cumsumNA.rainAllMonthly}
\alias{cumsumNA.rainAllAnnual}
\title{cumsumNA}
\usage{
cumsumNA(x, ...)

\method{cumsumNA}{numeric}(x)

\method{cumsumNA}{rainAll}(x, offset = 0)

\method{cumsumNA}{rainAllDaily}(x, offset = 0)

\method{cumsumNA}{rainAllMonthly}(x, offset = 0)

\method{cumsumNA}{rainAllAnnual}(x, offset = 0)
}
\arguments{
\item{x}{Raw rain gauge data or merged dataset}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Provides a cumulative sum of rain gauge data while excluding NA
values.
}
\examples{
cumsumNA(rain)
}
