% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{dailyAgg}
\alias{dailyAgg}
\alias{dailyAgg.flowLoad}
\alias{dailyAgg.rainLoad}
\alias{dailyAgg.stageLoad}
\alias{dailyAgg.rainAll}
\title{dailyAgg}
\usage{
dailyAgg(x, method = "mean", ...)

\method{dailyAgg}{flowLoad}(x, method = "mean", ...)

\method{dailyAgg}{rainLoad}(x, method = "mean", ...)

\method{dailyAgg}{stageLoad}(x, method = "mean", ...)

\method{dailyAgg}{rainAll}(x, method = "mean", ...)
}
\arguments{
\item{x}{Data generated in the HydroEnR package}

\item{method}{'mean', 'median', 'max', 'min', or 'sum'}

\item{...}{Other variables as required}
}
\value{

}
\description{
Aggregates sub daily time series into a daily resolution,
aggregations are carried out to the calendar day not a rolling time period.
}
\examples{
dailyAgg(Buildwas)
}
