% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Estimates.R
\name{Estimates}
\alias{Estimates}
\alias{Estimates.GEVPar}
\alias{Estimates.volGEVPar}
\alias{Estimates.GumbelPar}
\alias{Estimates.GenLogPar}
\alias{Estimates.GenParetoPar}
\title{Return Period Estimate}
\usage{
Estimates(x, q, RP, ppy, ...)

\method{Estimates}{GEVPar}(x, q = NULL, RP = 100)

\method{Estimates}{volGEVPar}(x, q = NULL, RP = 100)

\method{Estimates}{GumbelPar}(x, q = NULL, RP = 100)

\method{Estimates}{GenLogPar}(x, q = NULL, RP = 100)

\method{Estimates}{GenParetoPar}(x, q = NULL, RP = 100, ppy = 1)
}
\arguments{
\item{x}{Parameters defined from various other functions}

\item{q}{An observed flow value, set to NULL but use to find return period}

\item{RP}{Set to 100, return period that you wish to generate a flow estimate for}

\item{ppy}{Peaks per year, set as one for Gen Pareto only}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Generate a flow estimate based on a user defined return period.
Alternatively add an observed flow to provide an estimate of return period.
}
\examples{
Estimates(GEVParams(buildwas), RP = c(100,200))
Estimates(GumbelParams(buildwas))
Estimates(GenLogParams(buildwas_max), q = 650)
Estimates(GenParetoParams(buildwas), ppy =2)
}
