% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchCDs.R
\name{fetchCDs}
\alias{fetchCDs}
\title{Fetch CDs}
\usage{
fetchCDs(CD, peakCDs)
}
\arguments{
\item{CD}{can be numeric ID relating to peak flows database or a file path to catchment descriptor file .cd2, .cd3, or .xml}

\item{peakCDs}{Peak flow database, optional}
}
\value{

}
\description{
Fetch CDs
}
\examples{
#fetchCDs(2001, peakFlowDatabase)
}
