% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks.R
\name{find_peaks}
\alias{find_peaks}
\title{Find peaks with base R}
\usage{
find_peaks(x, m = 3)
}
\arguments{
\item{x}{Data set}

\item{m}{Window of which to identify peaks}
}
\value{

}
\description{
Find peaks with base R
}
\examples{
find_peaks(allington$Value, m = 5000)
}
