% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeaks.R
\name{findPeaks}
\alias{findPeaks}
\title{Find peaks with base R}
\usage{
findPeaks(x, m = 3)
}
\arguments{
\item{x}{Data set}

\item{m}{Window of which to identify peaks}
}
\value{

}
\description{
Find peaks with base R
}
\examples{
findPeaks(allington$Value, m = 5000)
}
