% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findPeaksCPP}
\alias{findPeaksCPP}
\title{Find peaks in hydrological data}
\usage{
findPeaksCPP(vY, m = 3L)
}
\arguments{
\item{vY}{Numeric vector of hydrological data}

\item{m}{Window over which to determine peaks}
}
\description{
Tested to be 6x faster, 37 us vs 207 us. This operation is done from 200x per layer. Original R function by Stas_G
}
