% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaugeProp.R
\name{gaugeProp}
\alias{gaugeProp}
\title{Gauge proportion}
\usage{
gaugeProp(coords, catchment)
}
\arguments{
\item{coords}{Coordinates of rain gauges set as a polygon shapefile}

\item{catchment}{Catchment shapefile}
}
\value{

}
\description{
Calculates the the proportion of catchment that the gauge occupies using Thiessen / Voronoi polygons
}
\examples{
#gaugeProp(coords, shape_t)
}
