% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params.R
\name{GenLogParams}
\alias{GenLogParams}
\title{Determine parameters of Generalised Logistic fit}
\usage{
GenLogParams(x, URBEXT2000 = NULL, DeUrb = FALSE, ...)
}
\arguments{
\item{x}{Dataset of HydroEnR classes or L-moments}

\item{URBEXT2000}{If you wish to add urban adjustment the L-moments}

\item{DeUrb}{Set to FALSE, set to TRUE should you want to deurbanise}

\item{...}{Additional parameters as required}

\item{RP}{Vector of required return periods}
}
\value{

}
\description{
Determine parameters of Generalised Logistic fit
}
\examples{
#GenLogParams(buildwas)
}
