% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMAX.R
\name{getAMAX}
\alias{getAMAX}
\alias{getAMAX.numeric}
\alias{getAMAX.HydroAggsmax}
\alias{getAMAX.flowLoad}
\alias{getAMAX.zoo}
\title{getAMAX}
\usage{
getAMAX(x, ...)

\method{getAMAX}{numeric}(x = flow, Date = date, ...)

\method{getAMAX}{HydroAggsmax}(x)

\method{getAMAX}{flowLoad}(x, ...)

\method{getAMAX}{zoo}(x, ...)
}
\arguments{
\item{x}{A dataset containing flow data, generated from HydroEnR functions}

\item{...}{Extra parameters if required}
}
\value{

}
\description{
Extract annual maximum peak data from various data sources.
}
\examples{
getAMAX(Flows = Buildwas$Value, Date = Buildwas$DateTime)
getAMAX(Buildwas_Analysis)
getAMAX(Buildwas)
rnrfa::get_ts(id = 2001, type = 'amax-flow') \%>\% getAMAX()
getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow'))
}
