% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{getCoords}
\alias{getCoords}
\title{getCoords}
\usage{
getCoords(...)
}
\arguments{
\item{...}{The files you wish to include}
}
\value{

}
\description{
This extracts the site name, Latitude, and Longitude data from the imported
WISKI metadata associated with each file. Outputs are a shapefile of class SF.
}
\examples{
getCoords(oundle, yelden, corby)
}
