% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getVAMAX.R
\name{getVAMAX}
\alias{getVAMAX}
\alias{getVAMAX.HydroAggssum}
\alias{getVAMAX.flowLoad}
\title{getVAMAX}
\usage{
getVAMAX(x, ...)

\method{getVAMAX}{HydroAggssum}(x)

\method{getVAMAX}{flowLoad}(x, interval = 0.25, rolling_aggregations = c(1, 2, 3, 4, 8, 24, 120))
}
\arguments{
\item{x}{A dataset containing flow volume data, generated from HydroEnR functions}

\item{...}{Extra parameters if required}

\item{interval}{Set as 0.25 to represent 15 minute data, for hourly change to 1 etc.}

\item{rolling_aggregations}{User defined periods of aggregation}

\item{method}{Locked to 'sum'}
}
\value{

}
\description{
Extract annual maximum volume data from various data sources.
}
\examples{
#getVAMAX(buildwas)
#getVAMAX(buildwas_analysis)
}
