% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthFactors.R
\name{growthFactors}
\alias{growthFactors}
\title{Calculate the growth factors for all methods}
\usage{
growthFactors(
  x,
  RP = c(2, 4, 10, 25, 50, 100, 200, 1000),
  ppy = 1,
  URBEXT2000 = NULL,
  DeUrb = FALSE,
  ...
)
}
\arguments{
\item{x}{Dataset of HydroEnR classes or L-moments}

\item{RP}{Vector of required return periods}

\item{ppy}{Peaks per year}

\item{URBEXT2000}{If you wish to add urban adjustment the L-moments}

\item{DeUrb}{Set to FALSE, set to TRUE should you want to deurbanise}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculate the growth factors for all methods
}
\examples{
# growthFactors(buildwas_analysis)
# growthFactors(Ls(buildwas))
}
