% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthVolFactors.R
\name{growthVolGEV}
\alias{growthVolGEV}
\alias{growthVolGEV.vLs}
\title{Volumetric Generalised Extreme Value Growth Factors}
\usage{
growthVolGEV(x, ...)

\method{growthVolGEV}{vLs}(
  x,
  RP = c(2, 4, 10, 25, 50, 100, 200, 1000),
  URBEXT2000 = NULL,
  DeUrb = FALSE,
  ...
)
}
\arguments{
\item{x}{Dataset of HydroEnR classes or L-moments}

\item{...}{Additional parameters as required}

\item{RP}{Vector of required return periods}

\item{URBEXT2000}{If you wish to add urban adjustment the L-moments}

\item{DeUrb}{Set to FALSE, set to TRUE should you want to deurbanise}
}
\value{

}
\description{
Volumetric Generalised Extreme Value Growth Factors
}
\examples{
# growthVolGEV.vLs(getVAMAX(buildwas, URBEXT2000 = 0.3))
}
