% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{hourlyAgg}
\alias{hourlyAgg}
\alias{hourlyAgg.flowLoad}
\alias{hourlyAgg.stageLoad}
\alias{hourlyAgg.rainLoad}
\alias{hourlyAgg.rainAll}
\title{hourlyAgg}
\usage{
hourlyAgg(x, method = "mean", ...)

\method{hourlyAgg}{flowLoad}(x, method = "mean", ...)

\method{hourlyAgg}{stageLoad}(x, method = "mean", ...)

\method{hourlyAgg}{rainLoad}(x, method = "mean", ...)

\method{hourlyAgg}{rainAll}(x, method = "mean", ...)
}
\arguments{
\item{x}{Data generated in the HydroEnR package}

\item{method}{'mean', 'median', 'max', 'min', or 'sum'}

\item{...}{Other variables as required}
}
\value{

}
\description{
Aggregates sub hourly time series into an hourly resolution,
aggregations are carried out to the calendar day not a rolling time period.
}
\examples{
hourlyAgg(Buildwas)
}
