% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{hydroAggregate}
\alias{hydroAggregate}
\title{Hydro Aggregate}
\usage{
hydroAggregate(
  dt,
  interval = 0.25,
  rolling_aggregations = c(1, 2, 3, 4, 8, 24, 120),
  method = "mean"
)
}
\arguments{
\item{dt}{flowLoad, rainLoad, or stageLoad data}

\item{interval}{Set as 0.25 to represent 15 minute data, for hourly change to 1 etc.}

\item{rolling_aggregations}{User defined periods of aggregation}

\item{method}{'min', 'max', 'mean', 'median', and 'sum' options available}
}
\value{

}
\description{
Hydro Aggregate
}
\details{
Wrapper function for the various aggregation methods employed in
HydroEnR
}
\examples{
#hydroAggregate(buildwas)
}
