% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAPI.R
\name{hydroAPI}
\alias{hydroAPI}
\title{Load hydrometric data off the EAs API}
\usage{
hydroAPI(
  WISKI = NULL,
  Parameter = NULL,
  Period = NULL,
  Type = "instantaneous",
  datapoints = "standard",
  UAT = FALSE
)
}
\arguments{
\item{WISKI}{WISKI ID for gauge}

\item{Parameter}{User dfined; 'flow', 'level', 'rain'}

\item{Period}{Userdefined 900 == 15min, 86400 == daily}

\item{Type}{Data type set as 'instantaneous', some sites will have different agregators applied}

\item{datapoints}{Set to standard that pulls up to 100,000 rows. If 'all' will pull up to 2,000,000 rows. 'earliest' returns first datapoint}

\item{UAT}{Set to FALSE, if TRUE will pull data from UAT environment}
}
\value{

}
\description{
Load hydrometric data off the EAs API
}
\examples{
hydroAPI(WISKI = 'SX87U053')
hydroAPI(WISKI = 'SX87U053', UAT = TRUE)

hydroAPI(WISKI = '188753', UAT = FALSE)
hydroAPI(WISKI = '188753', UAT = TRUE)
# Pull data
hydroAPI(WISKI = 'SX87U053',
         Parameter = 'flow',
         Period = 900,
         Type = 'instantaneous',
         datapoints = 'standard')

hydroAPI(WISKI = 'SX87U053',
         Parameter = 'level',
         Period = 900,
         Type = 'instantaneous',
         datapoints = 'standard',
         UAT = TRUE)

hydroAPI(WISKI = '188753',
         Parameter = 'rainfall',
         Period = 86400,
         Type = 'total',
         datapoints = 'standard',
         UAT = TRUE)

}
