% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{hydroYearAgg}
\alias{hydroYearAgg}
\alias{hydroYearAgg.flowLoad}
\alias{hydroYearAgg.rainLoad}
\alias{hydroYearAgg.stageLoad}
\title{hydroYearAgg}
\usage{
hydroYearAgg(x, method = "mean", ...)

\method{hydroYearAgg}{flowLoad}(x, method = "mean", ...)

\method{hydroYearAgg}{rainLoad}(x, method = "mean", ...)

\method{hydroYearAgg}{stageLoad}(x, method = "mean", ...)
}
\arguments{
\item{x}{Data generated in the HydroEnR package}

\item{method}{'mean', 'median', 'max', 'min', or 'sum'}

\item{...}{Other variables as required}
}
\value{

}
\description{
Aggregates sub annual time series into a hydrological year resolution,
aggregations are carried out to the calendar day not a rolling time period.
}
\examples{
hydroYearAgg(Buildwas)
}
