% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroYearDay.R
\name{HydroYearDay}
\alias{HydroYearDay}
\title{HydroYearDay}
\usage{
HydroYearDay(d, hy_cal = "oct_us_gb")
}
\arguments{
\item{d}{Array of dates of class Date}

\item{hy_cal}{Set to 'oct_us_gb' (USA and UK), but can also be 'sep_br'
(Brazil), 'apr_cl' (Chille).}
}
\value{

}
\description{
Creates fields for hydrological year and hydrological day.

Speed is slow, needs conversion to C++ in future iterrations.
}
\examples{
HydroYearrDay(Buildwas$GaugeData$DateTime, hy_cal = 'oct_us_gb')
}
