% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teesunPlot.R
\name{intersectPoly}
\alias{intersectPoly}
\title{Intersect of polygons}
\usage{
intersectPoly(voronoi, catchment, coords)
}
\arguments{
\item{voronoi}{Voronoi/Thiessen polygon}

\item{catchment}{Catchment shapefile}

\item{coords}{Coordinates of rain gauges set as a polygon shapefile}
}
\value{

}
\description{
Creates intersected polygons of catchment against voronoi polygons
}
\examples{
#intersectPoly(voronoi = teesun(coords), catchment = shapefile, coords = coordinates)
}
