% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L1.R
\name{L1}
\alias{L1}
\alias{L1.numeric}
\alias{L1.HydroAggsmax}
\alias{L1.HydroAMAX}
\alias{L1.flowLoad}
\alias{print.L1}
\title{L1 moment}
\usage{
L1(x, ...)

\method{L1}{numeric}(x, ...)

\method{L1}{HydroAggsmax}(x, ...)

\method{L1}{HydroAMAX}(x, ...)

\method{L1}{flowLoad}(x, ...)

\method{print}{L1}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the L1 moment from AMAX data. Can also convert time
series data imported via HydroEnR into L1 moments.
}
\examples{
data <- getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow'))
L1(data)
}
