% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L3.R
\name{L3}
\alias{L3}
\alias{L3.numeric}
\alias{L3.HydroAggsmax}
\alias{L3.HydroAMAX}
\alias{L3.flowLoad}
\alias{print.L3}
\title{L3 moment}
\usage{
L3(x, ...)

\method{L3}{numeric}(x, ...)

\method{L3}{HydroAggsmax}(x, ...)

\method{L3}{HydroAMAX}(x, ...)

\method{L3}{flowLoad}(x, ...)

\method{print}{L3}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the L3 moment from AMAX data. Can also convert time
series data imported via HydroEnR into L3 moments.
}
\examples{
L3(Buildwas)
L3(Buildwas_Analysis)
L3(getAMAX(Buildwas))
L3(Buildwas_Analysis$Hydro_year$Hydro_year_Max)
}
