% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L4.R
\name{L4}
\alias{L4}
\alias{L4.numeric}
\alias{L4.HydroAggsmax}
\alias{L4.HydroAMAX}
\alias{L4.fowLoad}
\alias{print.L4}
\title{L4 moment}
\usage{
L4(x, ...)

\method{L4}{numeric}(x, ...)

\method{L4}{HydroAggsmax}(x, ...)

\method{L4}{HydroAMAX}(x, ...)

\method{L4}{fowLoad}(x, ...)

\method{print}{L4}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the L4 moment from AMAX data. Can also convert time
series data imported via HydroEnR into L4 moments.
}
\examples{
L4(Buildwas)
L4(Buildwas_Analysis)
L4(getAMAX(Buildwas))
L4(Buildwas_Analysis$Hydro_year$Hydro_year_Max)
}
