% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCA.R
\name{LCA}
\alias{LCA}
\alias{LCA.numeric}
\alias{LCA.HydroAggsmax}
\alias{LCA.HydroAMAX}
\alias{LCA.flowLoad}
\alias{print.LCA}
\title{LCA moment}
\usage{
LCA(x, ...)

\method{LCA}{numeric}(x, ...)

\method{LCA}{HydroAggsmax}(x, ...)

\method{LCA}{HydroAMAX}(x, ...)

\method{LCA}{flowLoad}(x, ...)

\method{print}{LCA}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the LCA moment from AMAX data. Can also convert time
series data imported via HydroEnR into LCA moments.
}
\examples{
LCA(Buildwas)
LCA(Buildwas_Analysis)
LCA(getAMAX(Buildwas))
LCA(Buildwas_Analysis$Hydro_year$Hydro_year_Max)
}
