% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LComponents.R
\name{LComponents}
\alias{LComponents}
\alias{LComponents.numeric}
\alias{LComponents.HydroAggsmax}
\alias{LComponents.HydroAMAX}
\alias{LComponents.flowLoad}
\alias{print.LComponents}
\title{L components}
\usage{
LComponents(x, ...)

\method{LComponents}{numeric}(x, ...)

\method{LComponents}{HydroAggsmax}(x, ...)

\method{LComponents}{HydroAMAX}(x, ...)

\method{LComponents}{flowLoad}(x, ...)

\method{print}{LComponents}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the components of the L moment calculations from AMAX data. Can also convert time
series data imported via HydroEnR into LKur moments.
}
\examples{
data <- getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow'))
LComponents(data)
}
