% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LKur.R
\name{LKur}
\alias{LKur}
\alias{LKur.numeric}
\alias{LKur.HydroAggsmax}
\alias{LKur.HydroAMAX}
\alias{LKur.flowLoad}
\alias{print.LKur}
\title{LKur moment}
\usage{
LKur(x, ...)

\method{LKur}{numeric}(x, ...)

\method{LKur}{HydroAggsmax}(x, ...)

\method{LKur}{HydroAMAX}(x, ...)

\method{LKur}{flowLoad}(x, ...)

\method{print}{LKur}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the LKur moment from AMAX data. Can also convert time
series data imported via HydroEnR into LKur moments.
}
\examples{
LKur(Buildwas)
LKur(Buildwas_Analysis)
LKur(getAMAX(Buildwas))
LKur(Buildwas_Analysis$Hydro_year$Hydro_year_Max)
}
