% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAPI.R
\name{loadAPI}
\alias{loadAPI}
\title{Pull data from the Environment Agency's API}
\usage{
loadAPI(
  ID = NULL,
  measure = NULL,
  period = NULL,
  type = NULL,
  datapoints = "standard",
  from = NULL,
  to = NULL,
  lat = NULL,
  long = NULL,
  easting = NULL,
  northing = NULL,
  dist = NULL,
  obsProperty = NULL
)
}
\arguments{
\item{ID}{Use to specify a particular WISKI ID or for when downloading all
stations ' wiski' or 'nrfa' to filter results}

\item{measure}{Use this when exporting observations to select the available
parameter. Generally 'flow', 'level', or 'groundwater'}

\item{period}{This is so you can select the time steps available, generally
900 (15 min) or 86400 (daily)}

\item{type}{Selects the data type, instantaneous, min, max etc.}

\item{datapoints}{Allows you to decide where in the time series you are
pulling the data from}

\item{from}{First time step of choice}

\item{to}{Last time step of choice}

\item{lat}{Latitude coordinate for when searching by geographic area}

\item{long}{Longitude coordinate for when searching by geographic area}

\item{easting}{Easting coordinate for when searching by geographic area}

\item{northing}{Northing coordinate for when searching by geographic area}

\item{dist}{Distance in km for how far you wish to search}

\item{obsProperty}{Used to filter the stations when identifying sites.
Available metrics;
'waterFlow', 'waterLevel', 'rainfall', 'groundwaterLevel', 'ammonium',
'dissolved-oxygen', 'conductivity', 'ph', 'temperature', 'turbidity',
'nitrate', 'chlorophyll', 'salinity', 'bga', 'fdom'}
}
\value{

}
\description{
The loadAPI function is used to interogate the EAs API. Data for
all station can be pulled, or when specified to a single site various
datasets can be pulled.
}
\examples{
loadAPI()

loadAPI(easting = 378235 , northing = 276165, dist = 30, ID = 'nrfa')

loadAPI(ID = 'L1207')

loadAPI(ID = 'L1207',
measure = 'level',
period = 900,
type = 'instantaneous',
datapoints = 'earliest')

dt <- loadAPI(ID = 'L1207',
measure = 'level',
period = 900,
type = 'instantaneous',
datapoints = 'all')

with(dt, plot(value ~ dateTime,
type = 'l',
xlab = 'Time',
ylab = 'Stage (mAoD)'))
}
