% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPE.R
\name{loadPE}
\alias{loadPE}
\title{Load potential evaporation data}
\usage{
loadPE(link, skip = 0, meta_rows = 10, impute = TRUE)
}
\arguments{
\item{link}{Link to the specified file for import}

\item{skip}{Stet to zero, denotes the number off rows you wish to skip}

\item{meta_rows}{Set to 10, determines the amount of rows set as metadata}

\item{impute}{Set to TRUE, this fills in missing data using the
'forecast::na.interp()' function}
}
\value{

}
\description{
Load potential evaporation data
}
\details{
Loads PE data obtained from HYRAD. Missing data are imputed using
the 'na.interp()' function in the forecast package. For the seasonal series,
a robust STL decomposition is first computed. Then a linear interpolation is
applied to the seasonally adjusted data, and the seasonal component is added
back.
}
