% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadPeakCDs.R
\name{loadPeakCDs}
\alias{loadPeakCDs}
\title{Load CD data from NRFA}
\usage{
loadPeakCDs(local = NULL)
}
\arguments{
\item{local}{defaults as NULL. To load a local version of the peak flow database and a file link.}
}
\value{

}
\description{
Load CD data from NRFA
}
\examples{
loadPeakCDs() \%>\%
  head()
}
