% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadWISKI.R
\name{loadWISKI}
\alias{loadWISKI}
\title{loadWISKI}
\usage{
loadWISKI(
  link,
  skip = 0,
  meta_rows = 15,
  hydro_year = "oct_us_gb",
  cols = NULL
)
}
\arguments{
\item{link}{Link to the specified file for import}

\item{skip}{Stet to zero, denotes the number off rows you wish to skip}

\item{meta_rows}{Set to 15, determines the amount of rows set as metadata}

\item{hydro_year}{Set to 'oct_us_gb' (USA and UK), but can also be 'sep_br'
(Brazil),'apr_cl' (Chille).}

\item{cols}{Set to NULL, but can be set as a vector of column names. Note
the first 2 must be 'DateTime' and 'Value'}
}
\value{

}
\description{
Load data downloaded from WISKI, in the EA ascii format. Metadata
are preserved. Loaded data are in the form of a list, called either flowLoad,
rainLoad, or stageLoad. Data are in two parts MetaData and GaugeData.
}
\examples{
#csvfile <- 'C:/Users/jpayne05/Downloads/Wellingborough.csv'
#a <- loadWISKI(csvfile)
}
