% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ls.R
\name{Ls}
\alias{Ls}
\title{L moments}
\usage{
Ls(x, URBEXT2000 = NULL, DeUrb = FALSE, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the L moments from AMAX data. Can also convert time
series data imported via HydroEnR into LKur moments.
}
\examples{
data <- getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow'))
Ls(data)
}
