% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSKEW.R
\name{LSKEW}
\alias{LSKEW}
\alias{LSKEW.numeric}
\alias{LSKEW.HydroAggsmax}
\alias{LSKEW.HydroAMAX}
\alias{LSKEW.flowLoad}
\alias{print.LSkew}
\title{L-Skew moment}
\usage{
LSKEW(x, ...)

\method{LSKEW}{numeric}(x)

\method{LSKEW}{HydroAggsmax}(x)

\method{LSKEW}{HydroAMAX}(x)

\method{LSKEW}{flowLoad}(x)

\method{print}{LSkew}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the L-Skew moment from AMAX data. Can also convert time
series data imported via HydroEnR into L-Skew moments.
}
\examples{
LSKEW(Buildwas)
LSKEW(Buildwas_Analysis)
Urb(LSKEW(Buildwas_Analysis), 0.3)
LSKEW(Buildwas_Analysis$Hydro_year$HydroYear_Max) \%>\% Urb(0.3)
LSKEW(GetAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow')))
}
