% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objectives.R
\name{MAE}
\alias{MAE}
\title{Mean absolute error}
\usage{
MAE(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{Observed data series}

\item{y}{Estimated/modelled values}

\item{na.rm}{NA remove defaulted as FALSE}
}
\value{

}
\description{
Calculate MAE model diagnostic from an observed series,
x, and modelled series, y.
}
\examples{
MAE(x, y)
}
