% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Merge WISKI datasets}
\usage{
mergeData(...)
}
\arguments{
\item{...}{Multiple flowLoad, rainLoad, and /or stageLoad class data.}
}
\value{

}
\description{
Merges multiple time series data sets into a single data.table.
Pads shorter time series with NAs.
}
\examples{
rain <- mergeData(corby, brigstock, yelden, oundle, wellingborough, kingscliffe)
}
