% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaCollate.R
\name{metaCollate}
\alias{metaCollate}
\title{metaCollate}
\usage{
metaCollate(...)
}
\arguments{
\item{...}{WISKI deived datasets of classes; 'flowLoad', 'rrainLoad', and
'stageLoad'}
}
\value{

}
\description{
Collate the metadata of numerous WISKI datasets.
}
\examples{
# meta <- metaCollate(brigstock, corby, kingscliffe, oundle, wellingborough, yelden)
}
