% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregations.R
\name{monthlyAgg}
\alias{monthlyAgg}
\alias{monthlyAgg.flowLoad}
\alias{monthlyAgg.rainLoad}
\alias{monthlyAgg.stageLoad}
\alias{monthlyAgg.rainAll}
\title{monthlyAgg}
\usage{
monthlyAgg(x, method = "mean", ...)

\method{monthlyAgg}{flowLoad}(x, method = mean, ...)

\method{monthlyAgg}{rainLoad}(x, method = mean, ...)

\method{monthlyAgg}{stageLoad}(x, method = mean, ...)

\method{monthlyAgg}{rainAll}(x, method = "mean", ...)
}
\arguments{
\item{x}{Data generated in the HydroEnR package}

\item{method}{'mean', 'median', 'max', 'min', or 'sum'}

\item{...}{Other variables as required}
}
\value{

}
\description{
Aggregates sub monthly time series into a monthly resolution,
aggregations are carried out to the calendar day not a rolling time period.
}
\examples{
monthlyAgg(Buildwas)
}
