% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthPlot.R
\name{monthPlot}
\alias{monthPlot}
\alias{monthPlot.HydroAggs}
\alias{monthPlot.flowLoad}
\title{monthPlot}
\usage{
monthPlot(x, name, polar, method, snip, ...)

\method{monthPlot}{HydroAggs}(x, name = "Gauge", polar = FALSE, snip = NULL, method = "max", ...)

\method{monthPlot}{flowLoad}(x, name = "Gauge", polar = FALSE, snip = NULL, method = "max", ...)
}
\arguments{
\item{x}{Data set}

\item{name}{Gauge used}

\item{polar}{Set to FALSE, if TRUE plot will be circular}

\item{method}{Set as 'max' change to type of aggregation used}

\item{snip}{Used to determine where to cut time series using headTail function}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Creates a plot of monthly flows using the aggregations in the
hydroAggs function.
}
\examples{
monthplot(Buildwas_Analysis, method = 'sum', name = 'Buildwas', polar = FALSE)
monthplot(Buildwas_Analysis, name = 'Buildwas', polar = TRUE)
monthPlot(Buildwas)
}
