% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objectives.R
\name{Objectives}
\alias{Objectives}
\alias{R2}
\title{Model diagnostics}
\usage{
R2 (x, y, na.rm=FALSE)
RMSE (x, y, na.rm=FALSE)
MAE (x, y, na.rm=FALSE)
RMSEP (x, y, na.rm=FALSE)
MAEP (x, y, na.rm=FALSE)
}
\arguments{
\item{x}{Observed data series}

\item{y}{Estimated/modelled values}

\item{na.rm}{NA remove defaulted as FALSE}
}
\value{

}
\description{
Calculate model diagnostics from an observed series, x, and
modelled series, y.
}
\examples{
R2(x, y)
RMSE(x, y)
MAE(x, y)
RMSEP(x, y)
MAEP(x, y)
}
