% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{PDM}
\alias{PDM}
\title{Basic pareto based PDM in R}
\usage{
PDM(fc, Cmin, Cmax, b, P, PE, s, be, kg, St, bg, k1, k2, kb, qconst, area)
}
\arguments{
\item{fc}{Rainfac, adjust to apply multiplier against rainfall}

\item{Cmin}{Minimum soil storage}

\item{Cmax}{Maximum soil storage}
}
\description{
Basic pareto based PDM in R
}
