% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDet.R
\name{peakDet}
\alias{peakDet}
\title{peakDet}
\usage{
peakDet(v, delta, x = NULL)
}
\arguments{
\item{v}{a numeric vector where to search peaks.}

\item{delta}{numeric of length one; defining the local threshold for peak detection.}

\item{x}{a numeric vector the same length as v containing corresponding x-values for v.}
}
\value{

}
\description{
A peak detection algorithm for R; finds local maxima and minima.
Based on an algorithm for MATLAB by Eli Billauer (http://www.billauer.co.il/peakdet.html).

A list containing two data frames maxtab and mintab containing maxima and
minima. Data frames contain two columns with indices in v (or corresponding
values in x if provided) and values in v.
}
\examples{
x <- seq(0, 20, .05)
vals <- sin(x)
det <- peakdet(vals, .5, x)
det
plot(x = x, y = vals, frame = FALSE, xlab = 'Time', ylab = 'Stage', type = 'l');
points(det$maxtab$val~det$maxtab$pos, bg = 'red', pch = 21, col = 'black',
lwd = 0.9, cex = 1.5);
points(det$mintab$val~det$mintab$pos, bg = 'blue', pch = 21, col = 'black',
lwd = 0.9, cex = 1.5)
}
