% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotExtVol.R
\name{plotExtVol}
\alias{plotExtVol}
\title{Extreme Volume Plots}
\usage{
plotExtVol(
  x,
  duration = NULL,
  scaling = TRUE,
  filter = NULL,
  secX = TRUE,
  ppy = 1,
  URBEXT2000 = NULL,
  DeUrb = FALSE
)
}
\arguments{
\item{x}{HydroVAMAX Dataset}

\item{duration}{At the moment this function can only process one duration at a time. Select the column name}

\item{scaling}{Scale to QMED, set as TRUE}

\item{filter}{Set to null, which selects all. Use vector of distributions to select models; 'GEV', 'Gumbel', 'GenLog', and 'GenPareto'}

\item{secX}{Secondary X axis to highlight return periods}

\item{ppy}{Peaks per year used in the growth function equations}

\item{URBEXT2000}{Used in the growth function equations}

\item{DeUrb}{Used in the growth function equations}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Extreme Volume Plots
}
\examples{
amax <- getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow'))
plotExtVal(amax)
plotExtVal(amax, scaling = FALSE, filter = c('GEV', 'GenLog'), URBEXT2000 = 0.4)
}
