% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNA.R
\name{plotNA}
\alias{plotNA}
\alias{plotNA.rainAll}
\title{plotNA}
\usage{
plotNA(x)

\method{plotNA}{rainAll}(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Shows a proportions of missing data for each year.

Further updates will include flow and stage data.
}
\examples{
plotNA(rain)

}
