% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotQA.R
\name{plotQA}
\alias{plotQA}
\alias{plotQA.rainAll}
\alias{plotQA.rainAllDaily}
\alias{plotQA.rainAllMonthly}
\alias{plotQA.rainAllAnnual}
\alias{plotQA.cumulRain}
\title{plotQA}
\usage{
plotQA(x, ...)

\method{plotQA}{rainAll}(x, ...)

\method{plotQA}{rainAllDaily}(x, ...)

\method{plotQA}{rainAllMonthly}(x, ...)

\method{plotQA}{rainAllAnnual}(x, ...)

\method{plotQA}{cumulRain}(x, ...)
}
\arguments{
\item{x}{A data.table, list, or various other data type developed in the HydroEnR package}

\item{...}{Include extra graphing parameters}
}
\value{

}
\description{
Generic function for QA plots of HydroEnR objects.

Simple QA plots for various objects. Plotting adapts for the data type and accumulation.
}
\examples{
plotQA(b)
rain \%>\%
plotQA(zzz)
plotQA(cumsumNA.rainAll(window.rainAllDaily(zzz, start = '2005-06-12')))
plotQA(window(zzz, start = '2005-06-12'))
}
