% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QMED.R
\name{QMED}
\alias{QMED}
\alias{QMED.numeric}
\alias{QMED.HydroAggsmax}
\alias{QMED.HydroAMAX}
\alias{QMED.flowLoad}
\alias{print.QMED}
\title{Generate a QMED estimate}
\usage{
QMED(x, ...)

\method{QMED}{numeric}(x, ...)

\method{QMED}{HydroAggsmax}(x, ...)

\method{QMED}{HydroAMAX}(x, ...)

\method{QMED}{flowLoad}(x, ...)

\method{print}{QMED}(x, ...)
}
\arguments{
\item{x}{A series of flow data loaded into the environment via HydroEnR}

\item{...}{Additional parameters}
}
\value{

}
\description{
Calculate the QMED value of a series of flow data loaded into
the environment via HydroEnR.

Integrates a fix for the print function
}
\examples{
QMED(Buildwas)
QMED(Buildwas_Analysis)
}
