% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objectives.R
\name{R2}
\alias{R2}
\title{R-squared}
\usage{
R2(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{Observed data series}

\item{y}{Estimated/modelled values}

\item{na.rm}{NA remove defaulted as FALSE}
}
\value{

}
\description{
Calculate R-squared model diagnostic from an observed series,
x, and modelled series, y.
}
\examples{
R2(x, y)
}
