% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rainAvg.R
\name{rainAvg}
\alias{rainAvg}
\title{Merge Rain gauge data}
\usage{
rainAvg(gaugeProp, rainAll)
}
\arguments{
\item{gaugeProp}{Data table calculated with the gaugeProp() function}

\item{rainAll}{Data table of collated rain gauge data created with the mergeData() function}
}
\value{

}
\description{
Data from rain gauges are merged via this function using a weighted average
}
\examples{
#rainAvg(prop, rain)
}
