% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateOptimise.R
\name{rateOptimise}
\alias{rateOptimise}
\title{Optimise rating relationships on multiple limbs}
\usage{
rateOptimise(Discharge, Stage, control = NULL, ...)
}
\arguments{
\item{Discharge}{Flow series}

\item{Stage}{Level/stage series}

\item{control}{Change points in the rating}

\item{...}{Apply additional commands to the Non-linear Least Squares optimiser}
}
\value{

}
\description{
Optimises the rating relationships using the Levenberg-Marquardt algorithm
}
\examples{
#library(data.table)
#library(minpack.lm)
#gaugings <- data.frame(Stage = c(1.855, 2.109, 2.037, 1.972, 2.574, 1.748, 2.016),
 #                      Discharge = c(177.685, 240.898, 221.954, 205.55, 383.051, 154.061, 216.582))

# Writing data into an R dataframe:
#gaugings <- data.frame(stage=c(0.1374976,0.1354978,0.1547908,0.1547908,0.3047804,0.3022808,0.4138614,0.2901140,0.2320367,0.2841666,0.2524176,0.2185623,0.2666037,0.5372785,0.3462928,0.4944928),
 #                    discharge=c(0.004081598,0.004145695,0.005569669,0.005649690,0.024848787,0.025689611,0.059294500,0.020582500,0.009332000,0.021064750,0.016438000,0.010596000,0.017770950,0.139279063,0.040481681,0.092302000))

#Rating <- read.csv("C:/Users/jpayne05/OneDrive - Defra/Rating.csv")
#rateOptimise(Discharge = rate$Discharge, Stage = rate$Stage)
#rateOptimise(Discharge = Rating$Discharge, Stage = Rating$Stage, control = c(1.6, 2.0, 2.2))
#z <- rateOptimise(Discharge = Rating$Discharge, Stage = Rating$Stage, control = 1.6)
}
