% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rateOptimise.R
\name{ratingPlot}
\alias{ratingPlot}
\title{Plot rating objects}
\usage{
ratingPlot(fit, colours = NULL)
}
\arguments{
\item{fit}{Fittings derived from the rateOptimise function}

\item{Discharge}{Flow series}

\item{Stage}{Level/stage series}
}
\value{

}
\description{
Plot rating objects
}
\examples{
# z <- rateOptimise(Discharge = rate$Discharge, Stage = rate$Stage, control = 1.6)
# ratingPlot(fit = z, colours = c(5, 2, 3, 4))
}
