% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Objectives.R
\name{RMSEP}
\alias{RMSEP}
\title{Root mean square error percentage}
\usage{
RMSEP(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{Observed data series}

\item{y}{Estimated/modelled values}

\item{na.rm}{NA remove defaulted as FALSE}
}
\value{

}
\description{
Calculate RMSEP model diagnostic from an observed series,
x, and modelled series, y.
}
\examples{
RMSEP(x, y)
}
