% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{teeSun}
\alias{teeSun}
\title{teeSun}
\usage{
teeSun(x)
}
\arguments{
\item{x}{SF file with point data within.}
}
\value{

}
\description{
Creates a Thiessen/Voronoi polygon shapefile.
}
\examples{
teeSun(s)
}
