% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teesunPlot.R
\name{teesunPlot}
\alias{teesunPlot}
\title{teesunPlot}
\usage{
teesunPlot(coords, catchment)
}
\arguments{
\item{coords}{Coordinates of rain gauges set as a polygon shapefile}

\item{catchment}{Catchment shapefile}
}
\value{

}
\description{
Creates a map of a catchment with Thiessen / Vornoi polyggons overlayed
}
\examples{
# teesunPlot(coords = coordinates, catchment = catchment)

# z <- teesunPlot(coords = coords, catchment = shape_t)

#z \%>\% addProviderTiles(providers$Stamen.Toner)
#z \%>\% addProviderTiles(providers$CartoDB.Positron)
#z \%>\% addProviderTiles(providers$Esri.NatGeoWorldMap)
}
