% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uploadPeakFlow.R
\name{uploadPeakFlow}
\alias{uploadPeakFlow}
\title{Upload peak flow data from a local source}
\usage{
uploadPeakFlow(filepath)
}
\arguments{
\item{filepath}{file path to Peak Flow data folder}
}
\value{

}
\description{
Upload peak flow data from a local source
}
\examples{
#uploadPeakFlow(filepath = 'This/is/not/a/file.path')
}
