% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ls.R
\name{vLs}
\alias{vLs}
\title{Volumetric L moments}
\usage{
vLs(x, URBEXT2000 = NULL, DeUrb = FALSE, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the L moments from VAMAX data. Can also convert time
series data imported via HydroEnR into LKur moments.
}
\examples{
#vLs(data)
}
