% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VMED.R
\name{VMED}
\alias{VMED}
\alias{VMED.numeric}
\alias{VMED.HydroVAMAX}
\alias{VMED.HydroAggssum}
\alias{VMED.flowLoad}
\title{Generate a VMED estimate}
\usage{
VMED(x, ...)

\method{VMED}{numeric}(x)

\method{VMED}{HydroVAMAX}(x)

\method{VMED}{HydroAggssum}(x, interval = 0.25, rolling_aggregations = c(1, 2, 3, 4, 8, 24, 120))

\method{VMED}{flowLoad}(x, interval = 0.25, rolling_aggregations = c(1, 2, 3, 4, 8, 24, 120))
}
\arguments{
\item{x}{A series of flow data loaded into the environment via HydroEnR}

\item{...}{Extra parameters if required}

\item{interval}{Set as 0.25 to represent 15 minute data, for hourly change to 1 etc.}

\item{rolling_aggregations}{User defined periods of aggregation}

\item{method}{Locked to 'sum'}
}
\value{

}
\description{
Calculate the VMED value of a series of flow data loaded into
the environment via HydroEnR.
}
\examples{
#VMED(Buildwas)
#VMED(Buildwas_Analysis)
#VMED(getVAMAX(buildwas_analysis))
}
