% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volParams.R
\name{volGenLogParams}
\alias{volGenLogParams}
\alias{volGenLogParams.vLs}
\title{Determine parameters of Generalised Logistic fit for volumetric data}
\usage{
volGenLogParams(x, ...)

\method{volGenLogParams}{vLs}(x, URBEXT2000 = NULL, DeUrb = FALSE, ...)
}
\arguments{
\item{x}{Dataset of HydroEnR classes or L-moments}

\item{...}{Additional parameters as required}

\item{URBEXT2000}{If you wish to add urban adjustment the L-moments}

\item{DeUrb}{Set to FALSE, set to TRUE should you want to deurbanise}

\item{RP}{Vector of required return periods}
}
\value{

}
\description{
Determine parameters of Generalised Logistic fit for volumetric data
}
\examples{
#volGenLogParams(buildwas)
}
