% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volParams.R
\name{volGenParetoParams}
\alias{volGenParetoParams}
\alias{volGenParetoParams.vLs}
\title{Determine parameters of Generalised Pareto fit for volumetric data}
\usage{
volGenParetoParams(x, ...)

\method{volGenParetoParams}{vLs}(x, URBEXT2000 = NULL, DeUrb = FALSE, ...)
}
\arguments{
\item{x}{Dataset of HydroEnR classes or L-moments}

\item{...}{Additional parameters as required}

\item{URBEXT2000}{If you wish to add urban adjustment the L-moments}

\item{DeUrb}{Set to FALSE, set to TRUE should you want to deurbanise}

\item{RP}{Vector of required return periods}
}
\value{

}
\description{
Determine parameters of Generalised Pareto fit for volumetric data
}
\examples{
#volGenParetoParams(buildwas)
}
