% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volParams.R
\name{volGEVParams}
\alias{volGEVParams}
\alias{volGEVParams.vLs}
\title{Determine parameters of Generalised Extreme Value fit for volumetric data}
\usage{
volGEVParams(x, ...)

\method{volGEVParams}{vLs}(x, URBEXT2000 = NULL, DeUrb = FALSE, ...)
}
\arguments{
\item{x}{Dataset of HydroEnR classes or vL-moments}

\item{...}{Additional parameters as required}

\item{URBEXT2000}{If you wish to add urban adjustment the L-moments}

\item{DeUrb}{Set to FALSE, set to TRUE should you want to deurbanise}

\item{RP}{Vector of required return periods}
}
\value{

}
\description{
Determine parameters of Generalised Extreme Value fit for volumetric data
}
\examples{
#volGEVParams(buildwas)
}
