% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volParams.R
\name{volGumbelParams}
\alias{volGumbelParams}
\alias{volGumbelParams.vLs}
\title{Determine parameters of Gumbel fit for volumetric data}
\usage{
volGumbelParams(x, ...)

\method{volGumbelParams}{vLs}(x, URBEXT2000 = NULL, DeUrb = FALSE, ...)
}
\arguments{
\item{x}{Dataset of HydroEnR classes or vL-moments}

\item{...}{Additional parameters as required}

\item{URBEXT2000}{If you wish to add urban adjustment the L-moments}

\item{DeUrb}{Set to FALSE, set to TRUE should you want to deurbanise}

\item{RP}{Vector of required return periods}
}
\value{

}
\description{
Determine parameters of Gumbel fit for volumetric data
}
\examples{
#volGumbelParams(buildwas)
}
