% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/window.R
\name{window}
\alias{window}
\alias{window.rainAll}
\alias{window.rainAllDaily}
\title{Time (Series) Windows}
\usage{
window(x, start, end)

\method{window}{rainAll}(x, start = NULL, end = NULL)

\method{window}{rainAllDaily}(x, start = NULL, end = NULL)
}
\arguments{
\item{x}{a data table or other input table}

\item{start}{Start date and time. If NULL it will default to the start of the
series.}

\item{end}{End date and time. If NULL it will default to the end of the
series}
}
\value{

}
\description{
window is a generic function which extracts the subset of the
object x observed between the times start and end. If a frequency is specified, the
series is then re-sampled at the new frequency.

Currently this is only valid with merged rainfall data tables.

Future updates will include application to other data sets
}
\examples{
window(rain, start = '1988-06-12 23:30')
}
