% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zataTable.R
\name{zataTable}
\alias{zataTable}
\alias{zataTable.zoo}
\title{zataTable}
\usage{
zataTable(x, ...)

\method{zataTable}{zoo}(x, index.name = "Date")
}
\arguments{
\item{x}{data in zoo format}

\item{...}{Extra parameters if requirred}
}
\value{

}
\description{
Converts AMAX data taken from NRFA from type zoo to data.table.
Maintaining consistency in package.
}
\examples{
rnrfa::get_ts(id = 2001, type = 'amax-flow') \%>\% zataTable()
}
