% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2.R
\name{L2}
\alias{L2}
\alias{L2.numeric}
\alias{L2.HydroAggsmax}
\alias{L2.HydroAMAX}
\alias{L2.flowLoad}
\alias{L2.zoo}
\alias{print.L2}
\title{L2 moment}
\usage{
L2(x, ...)

\method{L2}{numeric}(x, ...)

\method{L2}{HydroAggsmax}(x, ...)

\method{L2}{HydroAMAX}(x, ...)

\method{L2}{flowLoad}(x, ...)

\method{L2}{flowLoad}(x, ...)

\method{L2}{zoo}(x, ...)

\method{print}{L2}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the L2 moment from AMAX data. Can also convert time
series data imported via HydroEnR into L2 moments.
}
\examples{
data <- getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow'))
L2(data)
}
