% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCA.R
\name{LCA}
\alias{LCA}
\alias{LCA.numeric}
\alias{LCA.HydroAggsmax}
\alias{LCA.HydroAMAX}
\alias{LCA.flowLoad}
\alias{LCA.zoo}
\title{LCA moment}
\usage{
LCA(x, ...)

\method{LCA}{numeric}(x, ...)

\method{LCA}{HydroAggsmax}(x, ...)

\method{LCA}{HydroAMAX}(x, ...)

\method{LCA}{flowLoad}(x, ...)

\method{LCA}{zoo}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the LCA moment from AMAX data. Can also convert time
series data imported via HydroEnR into LCA moments.
}
\examples{
data <- getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow'))
LCA(data)
}
