% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCV.R
\name{LCV}
\alias{LCV}
\alias{LCV.numeric}
\alias{LCV.HydroAggsmax}
\alias{LCV.HydroAMAX}
\alias{LCV.flowLoad}
\alias{print.Lcv}
\title{LCV moment}
\usage{
LCV(x, ...)

\method{LCV}{numeric}(x)

\method{LCV}{HydroAggsmax}(x)

\method{LCV}{HydroAMAX}(x)

\method{LCV}{flowLoad}(x)

\method{print}{Lcv}(x, ...)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the LCV moment from AMAX data. Can also convert time
series data imported via HydroEnR into LCV moments.
}
\examples{
LCV(Buildwas)
LCV(Buildwas_Analysis)
LCV(Buildwas_Analysis) \%>\% Urb(0.3)
LCV(Buildwas_Analysis) \%>\% urb(0.3)
LCV(Buildwas_Analysis$Hydro_year$Hydro_year_Max)
LCV(getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow')))
}
