% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSKEW.R
\name{LSKEW}
\alias{LSKEW}
\alias{LSKEW.numeric}
\alias{LSKEW.HydroAggsmax}
\alias{LSKEW.HydroAMAX}
\alias{LSKEW.flowLoad}
\alias{LSKEW.zoo}
\title{L-Skew moment}
\usage{
LSKEW(x, ...)

\method{LSKEW}{numeric}(x)

\method{LSKEW}{HydroAggsmax}(x)

\method{LSKEW}{HydroAMAX}(x)

\method{LSKEW}{flowLoad}(x)

\method{LSKEW}{zoo}(x)
}
\arguments{
\item{x}{AMAX or flow time series}

\item{...}{Additional parameters as required}
}
\value{

}
\description{
Calculates the L-Skew moment from AMAX data. Can also convert time
series data imported via HydroEnR into L-Skew moments.
}
\examples{
data <- getAMAX(rnrfa::get_ts(id = 2001, type = 'amax-flow'))
LSKEW(data)
}
