% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.R
\name{teeSun}
\alias{teeSun}
\title{teeSun}
\usage{
teeSun(x, catchment = NULL)
}
\arguments{
\item{x}{SF file with point data within.}

\item{catchment}{Set as null, optional bounding box polygon derived from catchment polygon}
}
\value{

}
\description{
Creates a Thiessen/Voronoi polygon shapefile.
}
\examples{
teeSun(s)
}
